+++ commands/src/apply/mod.rs
@@ -1,3 +1,2 @@
 pub mod init;
 pub mod lockfile;
-pub mod report;